% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-mappingV1.R
\name{mappingV1}
\alias{mappingV1}
\title{Backend V1 of the mapping API}
\usage{
mappingV1(mappingId = "IsoMemo")
}
\arguments{
\item{mappingId}{id to select one of the mappings, e.g. "IsoMemo"}
}
\description{
Backend V1 of the mapping API
}
