% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-wrapCall.R
\name{wrapCall}
\alias{wrapCall}
\title{Wrap call to API endpoints}
\usage{
wrapCall(expr)
}
\arguments{
\item{expr}{(expression) a call to the API endpoint provided by the package.}
}
\description{
This is overrides most of plumbers error handling mechanism and allows for
custom status flags and messages.
}
