% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dbsourcesV1.R
\name{dbsourcesV1}
\alias{dbsourcesV1}
\title{Backend V1 of the dbsources API}
\usage{
dbsourcesV1(mappingId = "IsoMemo")
}
\arguments{
\item{mappingId}{id to select one of the mappings, e.g. "IsoMemo"}
}
\description{
Backend V1 of the dbsources API
}
