% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-isoDataV1.R
\name{isoDataV1}
\alias{isoDataV1}
\title{Backend V1 of the API}
\usage{
isoDataV1(dbsource, category, field)
}
\arguments{
\item{dbsource}{(character) filter for database proider}

\item{category}{(character) filter for field category}

\item{field}{(character) filter for fields}
}
\description{
Backend V1 of the API
}
